moved all tests under  & fixed a test
